\name{GetCommonExtent}
\alias{GetCommonExtent}
\title{get the common extent of different rasterLayers}
\description{The function computes the minimum common extent of RasterLayers.}
\usage{GetCommonExtent(raster.l)}
\arguments{
  \item{raster.l}{a list of rasterLayers}
}

\value{The function returns an object of class extent}

\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}




\examples{
#
}
