\name{ModelDataComp-package}
\alias{ModelDataComp-package}
\alias{ModelDataComp}
\docType{package}
\title{Model-Data Comparison}
}
\description{Metrics and plots for model-data comparison. The package provides functions to 1) compute model performance metrics (function ObjFct), to 2) fit empirical models to data (function MultiFit), and to 3) make model-data comparison plots (functions ScatterPlot, plot.ObjFct with barplots and scatterplots of objective functions, TaylorPlot, WollMilchSauPlot to plot means, distributions and correlations between observations and models).}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ModelDataComp}
\packageIndices{ModelDataComp}
}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}

Maintainer: \packageMaintainer{ModelDataComp}
}

\keyword{ package }


