\name{print.ObjFct}
\alias{print.ObjFct}
\title{Print objective functions}
\description{Prints objective functions}
\usage{\method{print}{ObjFct}(x, ...)}
\arguments{
  \item{x}{an object of class \code{\link{ObjFct}}}
  \item{\dots}{further arguments passed to or from other methods}
}
\details{Prints an object of class \code{\link{ObjFct}}.}

\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{ObjFct}}}

