\name{ModelDataComp-package}
\alias{ModelDataComp-package}
\alias{ModelDataComp}
\docType{package}
\title{Model-Data Comparison}
}
\description{Metrics and plots for model-data comparison. The package provides functions to 1) compute objective functions (model performance metrics), to 2) fit empirical models to data (currently random forest and multiple logistic functions), and to 3) make model-data comparison plots (barplots and scatterplots of objective functions, Taylor diagrams, WollMilchSauPlot).}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ModelDataComp}
\packageIndices{ModelDataComp}
}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}

Maintainer: \packageMaintainer{ModelDataComp}
}

\keyword{ package }


